+++ elmyra.ip.access.epo/CHANGES.rst
@@ -5,6 +5,7 @@ elmyra.ip.access.epo CHANGES
 development
 ===========
 - ui, parameters: add button to popover for opening permalinks, tune urlcleaner regex
+- ui: added “ZeroClipboard 2.1.2”
 
 0.23.4
 ======