+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,9 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+
+0.24.0
+======
 - ui, parameters: add button to popover for opening permalinks, tune urlcleaner regex
 - ui: added “ZeroClipboard 2.1.2”
 - ui: add button to copy permalink to clipboard